Fix source maps for rollup builds
When substituting __SAPPER_CSS_PLACEHOLDER:foo__, tighten the regexp
to require matching quotation marks and to not match too much if
there's an error.

Also, make the replacement logic smart about recognizing when we're
matching JavaScript code that's been quoted (e.g., in source maps),
and to generate a substitution that's still valid in that context.

Fixes #808.